experimental/requirements.txt
PyBluez==0.23
pkg-config
libboost-python-dev
libboost-thread-dev
libbluetooth-dev >= 4.101
libglib2.0-dev